fix sp500 upload
