+++ src/main/java/ru/investbook/service/Sp500OfficialService.java
@@ -1,0 +1,92 @@
+/*
+ * InvestBook
+ * Copyright (C) 2022  Spacious Team <spacious-team@ya.ru>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package ru.investbook.service;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.poi.hssf.usermodel.HSSFWorkbook;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.spacious_team.table_wrapper.api.PatternTableColumn;
+import org.spacious_team.table_wrapper.api.TableColumn;
+import org.spacious_team.table_wrapper.api.TableHeaderColumn;
+import org.spacious_team.table_wrapper.api.TableRow;
+import org.spacious_team.table_wrapper.excel.ExcelSheet;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.investbook.entity.StockMarketIndexEntity;
+import ru.investbook.repository.StockMarketIndexRepository;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.time.Duration;
+import java.util.Objects;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class Sp500OfficialService implements Sp500Service {
+    private final StockMarketIndexRepository stockMarketIndexRepository;
+
+    @Transactional
+    @Override
+    public void update(InputStream sp500ExcelFile) {
+        try {
+            long t0 = System.nanoTime();
+            updateBy(sp500ExcelFile);
+            log.info("Индекс S&P 500 обновлен за {}", Duration.ofNanos(System.nanoTime() - t0));
+        } catch (Exception e) {
+            throw new RuntimeException("Не смог обновить значения индекса S&P 500", e);
+        }
+    }
+
+    private void updateBy(InputStream inputStream) throws IOException {
+        Workbook book = new HSSFWorkbook(inputStream);
+        new ExcelSheet(book.getSheetAt(0))
+                .createNameless("Effective date", TableHeader.class)
+                .stream()
+                .filter(Objects::nonNull)
+                .map(Sp500OfficialService::getIndexValue)
+                .forEach(this::save);
+    }
+
+    private static StockMarketIndexEntity getIndexValue(TableRow row) {
+        return StockMarketIndexEntity.builder()
+                .date(row.getLocalDateTimeCellValue(TableHeader.DATE).toLocalDate())
+                .sp500(row.getBigDecimalCellValue(TableHeader.VALUE))
+                .build();
+    }
+
+    private void save(StockMarketIndexEntity entity) {
+        try {
+            stockMarketIndexRepository.save(entity);
+        } catch (Exception e) {
+            log.debug("Ошибка сохранения {}, может быть запись уже существует?", entity);
+        }
+    }
+
+    @RequiredArgsConstructor
+    private enum TableHeader implements TableHeaderColumn {
+        DATE(PatternTableColumn.of("Effective date")),
+        VALUE(PatternTableColumn.of("S&P 500"));
+
+        @Getter
+        private final TableColumn column;
+    }
+}
